refactor: 당첨액과 구입 금액 데이터 타입 수정

+++ src/main/java/lotto/domain/LottoProfit.java
@@ -22,8 +22,8 @@ public class LottoProfit {
     }
 
     public void createLottoProfit(LottoPrize lottoPrize, LottoPurchase lottoPurchase) {
-        int totalWinningAmount = createTotalWinningAmount(lottoPrize);
-        int totalPurchaseAmount = createTotalPurchaseAmount(lottoPurchase);
+        double totalWinningAmount = createTotalWinningAmount(lottoPrize);
+        double totalPurchaseAmount = createTotalPurchaseAmount(lottoPurchase);
     }
 
     public int createTotalWinningAmount(LottoPrize lottoPrize) {
